Flowtest: Moving flowtest_helper under ocs heirarchy
